-moved Loader styles + variables to teams theme
