+++ packages/react/src/themes/base/componentVariables.ts
@@ -3,6 +3,4 @@ export { default as FlexItem } from './components/Flex/flexItemVariables'
 
 export { default as Icon } from './components/Icon/iconVariables'
 
-export { default as Loader } from './components/Loader/loaderVariables'
-
 export { default as Grid } from './components/Grid/gridVariables'
