+++ packages/react/src/themes/teams/components/Loader/loaderStyles.ts
@@ -1,22 +1,30 @@
+import { FlexDirectionProperty } from 'csstype'
 import { LoaderProps } from '../../../../components/Loader/Loader'
 import { ComponentStyleFunctionParam, ICSSInJSStyle } from '../../../types'
 import { LoaderVariables } from './loaderVariables'
+import { pxToRem } from '../../../../lib'
+import { ObjectOf } from '../../../../types'
+
+const rootFlexDirections: ObjectOf<FlexDirectionProperty> = {
+  above: 'column-reverse',
+  below: 'column',
+  start: 'row-reverse',
+  end: 'row',
+}
 
 export default {
+  root: ({
+    props: p,
+  }: ComponentStyleFunctionParam<LoaderProps, LoaderVariables>): ICSSInJSStyle => ({
+    alignItems: 'center',
+    display: p.inline ? 'inline-flex' : 'flex',
+    justifyContent: 'center',
+    flexDirection: rootFlexDirections[p.labelPosition],
+  }),
   indicator: ({
     props: p,
     variables: v,
   }: ComponentStyleFunctionParam<LoaderProps, LoaderVariables>): ICSSInJSStyle => ({
-    // Reset existing styles from base theme
-    animationName: 'none',
-    animationDuration: 'unset',
-    animationIterationCount: 'unset',
-    animationTimingFunction: 'unset',
-    borderColor: 'transparent',
-    borderRadius: 0,
-    borderStyle: 'none',
-    borderWidth: 0,
-
     height: v.containerHeights[p.size],
     width: v.containerWidths[p.size],
     overflow: 'hidden',
@@ -73,5 +81,8 @@ export default {
         width: v.svgWidths[p.size],
       },
     }
+  },
+  label: {
+    margin: pxToRem(10),
   },
 }
